+++ packages/core/src/main/java/dev/synapse/core/logging/LogLevel.java
@@ -1,0 +1,9 @@
+package dev.synapse.core.logging;
+
+public enum LogLevel {
+    DEBUG,
+    INFO,
+    WARN,
+    ERROR,
+    FATAL
+}